requirements.txt
opencv-python>=4.6
mediapipe==0.10.11
numpy>=1.22
matplotlib>=3.7
scikit-learn>=1.1
scikit-image>=0.19
torch==2.0.1
torchvision==0.15.2
Pillow>=10.0
PyQt5==5.15.4
pyttsx3>=2.90
tensorflow==2.13.1
protobuf==3.20.3
